Completed the Player class and its subclasses WhitePlayer and BlackPlayer.
Implemented the various functions in these classes that define the behavior of the player.
Also introduced a new class MoveTransition that will be used to represent the transition from one board state to another after a move is made.
The Move file now contains the MoveType enum to represent the type of move made.